rewriting submit js to handle sugars or context with regex and rewriting delete to not send a request everytime by a conditional on the id

+++ honey/public/javascripts/delete.js
@@ -5,27 +5,29 @@ table.addEventListener("click", handleRemove);
 let idForRemoval, action, removeRecord, recordForRemoval, response, tableRow;
 
 async function handleRemove(mouseEvent) {
-  try {
-    idForRemoval = mouseEvent.target.id;
+  idForRemoval = mouseEvent.target.id;
+  if (idForRemoval) {
+    try {
+      action = "/sugars/" + idForRemoval;
 
-    action = "/sugars/" + idForRemoval;
+      response = await fetch(action, {
+        method: "DELETE",
+        //   headers: { "Content-Type": "application/json" },
 
-    response = await fetch(action, {
-      method: "DELETE",
-      //   headers: { "Content-Type": "application/json" },
+        //   body: JSON.stringify(payload),
+      });
+    } catch (error) {
+      //remove element tr from DOM
+      console.log(error);
+    }
 
-      //   body: JSON.stringify(payload),
-    });
-    //remove element tr from DOM
-  } catch (error) {
-    console.log(error);
+    recordForRemoval = mouseEvent.target.parentElement;
+
+    response.status === 200
+      ? recordForRemoval.remove()
+      : console.log("resource not available");
   }
 
-  recordForRemoval = mouseEvent.target.parentElement;
-
-  response.status === 200
-    ? recordForRemoval.remove()
-    : console.log("resource not available");
   //   removeRecord = response ? true : false;
 
   //   if (removeRecord) {
